feat: 🎸 create BoardGameCafe, CafeGameList files
BoardGameCafe와 CafeGameList의 DTO, Service, Controller 생성 및 Swagger 설정

+++ yogidice/src/main/java/com/specialization/yogidice/dto/request/BoardGameRequest.java
@@ -25,9 +25,9 @@ public class BoardGameRequest {
     @NotNull
     private double rating;
 
-    private int minPlayers;
+    private byte minPlayers;
 
-    private int maxPlayers;
+    private byte maxPlayers;
 
     private String playingTime;
